requirements.txt
# Core data science libraries
pandas
pandas-stubs
numpy
scikit-learn

# Visualization
matplotlib
seaborn
plotly  # For interactive 3D visualizations
types-seaborn

# Specific algorithms
mlxtend  # For Apriori algorithm (Market Basket Analysis)

# Configuration
python-dotenv  # For loading environment variables from .env
